dev_requirements.txt
pip >= 20.0.0
flake8
pytest
pytest-django
